The script that likes to say yes

+++ setup/common/container-image-setup.sh
@@ -25,6 +25,77 @@ then
   exit 1
 fi
 
+
+# Get the OCIR locations
+echo "Locating repo names"
+OCIR_STOCKMANAGER_NAME=`oci artifacts  container repository get  --repository-id $OCIR_STOCKMANAGER_OCID | jq -r '.data."display-name"'`
+OCIR_STOREFRONT_NAME=`oci artifacts  container repository get  --repository-id $OCIR_STOREFRONT_OCID | jq -r '.data."display-name"'`
+
+
+echo "Checking for existing images"
+IMAGE_STOCKMANAGER_V001_OCID=`oci artifacts container image list --compartment-id $COMPARTMENT_OCID --display-name $OCIR_STOCKMANAGER_NAME:0.0.1 | jq -j ".data.items[0].id"`
+if [ -z $IMAGE_STOCKMANAGER_V001_OCID ]
+then
+  IMAGE_STOCKMANAGER_V001_OCID="null"
+fi
+IMAGE_STOCKMANAGER_V002_OCID=`oci artifacts container image list --compartment-id $COMPARTMENT_OCID --display-name $OCIR_STOCKMANAGER_NAME:0.0.2 | jq -j ".data.items[0].id"`
+if [ -z $IMAGE_STOCKMANAGER_V002_OCID ]
+then
+  IMAGE_STOCKMANAGER_V002_OCID="null"
+fi
+IMAGE_STOREFRONT_V001_OCID=`oci artifacts container image list --compartment-id $COMPARTMENT_OCID --display-name $OCIR_STOREFRONT_NAME:0.0.1 | jq -j ".data.items[0].id"`
+if [ -z $IMAGE_STOREFRONT_V001_OCID ]
+then
+  IMAGE_STOREFRONT_V001_OCID="null"
+fi
+IMAGE_STOREFRONT_V002_OCID=`oci artifacts container image list --compartment-id $COMPARTMENT_OCID --display-name $OCIR_STOREFRONT_NAME:0.0.2 | jq -j ".data.items[0].id"`
+if [ -z $IMAGE_STOREFRONT_V002_OCID ]
+then
+  IMAGE_STOREFRONT_V002_OCID="null"
+fi
+
+# if we have OCID's for all of the images no need to continue
+DO_BUILDS="false"
+if [ "$IMAGE_STOCKMANAGER_V001_OCID" = "null" ]
+then
+  echo "Missing stockmanager v0.0.1 image, build required"
+  DO_BUILDS="true"
+else
+  echo "Located image for stockmanager v0.0.1 image"
+fi
+
+if [ "$IMAGE_STOCKMANAGER_V002_OCID" = "null" ]
+then
+  echo "Missing stockmanager v0.0.2 image, build required"
+  DO_BUILDS="true"
+else
+  echo "Located image for stockmanager v0.0.2 image"
+fi
+
+
+if [ "$IMAGE_STOREFRONT_V001_OCID" = "null" ]
+then
+  echo "Missing storefront v0.0.1 image, build required"
+  DO_BUILDS="true"
+else
+  echo "Located image for storefront v0.0.1 image"
+fi
+
+
+if [ "$IMAGE_STOREFRONT_V002_OCID" = "null" ]
+then
+  echo "Missing storefront v0.0.2 image, build required"
+  DO_BUILDS="true"
+else
+  echo "Located image for storefront v0.0.2 image"
+fi
+
+if [ "$DO_BUILDS" = "false" ]
+then
+  echo "Found existing images for both storefront and stockmanager v0.0.1 and v0.0.2, no point in rebuilding"
+  echo "If you need to rebuild them then please destroy the existing images and re-run this script"
+  exit 0
+fi
 SCRIPTS_DIR=`pwd`
 
 
@@ -33,21 +104,32 @@ JAVA_LOCATION=https://download.oracle.com/java/17/latest/jdk-17_linux-x64_bin.ta
 # Old version of Java
 #JAVA_LOCATION=https://download.java.net/openjdk/jdk11/ri/openjdk-11+28_linux-x64_bin.tar.gz
 
+DEV_REL_GITHUB=https://github.com/oracle-devrel
+
 STOCKMANAGER_GIT_NAME=cloudnative-helidon-stockmanager
 STOREFRONT_GIT_NAME=cloudnative-helidon-storefront
 
-STOCKMANAGER_GIT_REPO=https://github.com/oracle-devrel/"$STOCKMANAGER_GIT_NAME".git
-STOREFRONT_GIT_REPO=https://github.com/oracle-devrel/"$STOREFRONT_GIT_NAME".git
+STOCKMANAGER_GIT_REPO="$DEV_REL_GITHUB"/"$STOCKMANAGER_GIT_NAME".git
+STOREFRONT_GIT_REPO="$DEV_REL_GITHUB"/"$STOREFRONT_GIT_NAME".git
 
 STOCKMANAGER_LOCATION_IN_REPO=helidon-stockmanager-full
 STOREFRONT_LOCATION_IN_REPO=helidon-storefront-full
 
-echo Removing any old directories
+echo "Removing any old directories"
 cd $HOME
-mkdir -p $WORK_DIR
-rm -rf $WORK_DIR
 
-echo About to install Java into $WORK_DIR from $JAVA_LOCATION
+if [ -d $WORK_DIR ]
+then
+  echo "Image build working directory $WORK_DIR exists, deleting"
+  rm -rf $WORK_DIR
+fi
+if [ -e $WORK_DIR ]
+then
+  echo "A file named the same as the image build working directory ($WORK_DIR) exists, deleting"
+  rm $WORK_DIR
+fi
+
+echo "About to install Java into $WORK_DIR from $JAVA_LOCATION"
 mkdir $WORK_DIR
 cd $WORK_DIR
 echo Downloading JDK
@@ -73,10 +155,6 @@ OBJECT_STORAGE_NAMESPACE=`oci os ns get | jq -j '.data'`
 
 echo "Building and pushing stockmanager images"
 
-# Get the OCIR location
-
-OCIR_STOCKMANAGER_NAME=`oci artifacts  container repository get  --repository-id $OCIR_STOCKMANAGER_OCID | jq -r '.data."display-name"'`
-
 cd $WORK_DIR/"$STOCKMANAGER_GIT_NAME"
 
 cd $STOCKMANAGER_LOCATION_IN_REPO
@@ -85,9 +163,21 @@ cd $STOCKMANAGER_LOCATION_IN_REPO
 # update the repo location
 echo "REPO=$OCIR_STOCKMANAGER_LOCATION/$OBJECT_STORAGE_NAMESPACE/$OCIR_STOCKMANAGER_NAME" > repoStockmanagerConfig.sh
 
-# build the images and push them
-bash buildStockmanagerPushToRepo.sh
-bash buildStockmanagerV0.0.2PushToRepo.sh
+# build the images and push them if needed
+if [ "$IMAGE_STOCKMANAGER_V001_OCID" = "null" ]
+then
+  bash buildStockmanagerPushToRepo.sh
+else 
+  echo "Located a Stockmanager v 0.0.1 image, reusing it"
+fi
+
+
+if [ "$IMAGE_STOCKMANAGER_V002_OCID" = "null" ]
+then
+  bash buildStockmanagerV0.0.2PushToRepo.sh
+else 
+  echo "Located a Stockmanager v 0.0.2 image, reusing it"
+fi
 
 cd $SCRIPTS_DIR
 
@@ -96,21 +186,27 @@ bash stockmanager-deployment-update.sh set $OCIR_STOCKMANAGER_LOCATION $OBJECT_S
 
 echo "Building and pushing storefront images"
 
-# Get the OCIR location
-
-OCIR_STOREFRONT_NAME=`oci artifacts  container repository get  --repository-id $OCIR_STOREFRONT_OCID | jq -r '.data."display-name"'`
-
 cd $WORK_DIR/"$STOREFRONT_GIT_NAME"
 
 cd $STOREFRONT_LOCATION_IN_REPO
-
 
 # update the repo location
 echo "REPO=$OCIR_STOREFRONT_LOCATION/$OBJECT_STORAGE_NAMESPACE/$OCIR_STOREFRONT_NAME" > repoStorefrontConfig.sh
 
 # build the images and push them
-bash buildStorefrontPushToRepo.sh
-bash buildStorefrontV0.0.2PushToRepo.sh
+if [ "$IMAGE_STOREFRONT_V001_OCID" = "null" ]
+then
+  bash buildStorefrontPushToRepo.sh
+else 
+  echo "Located a Storefront v 0.0.1 image, reusing it"
+fi
+
+if [ "$IMAGE_STOREFRONT_V002_OCID" = "null" ]
+then
+  bash buildStorefrontV0.0.2PushToRepo.sh
+else 
+  echo "Located a Storefront v 0.0.2 image, reusing it"
+fi
 
 cd $SCRIPTS_DIR
 bash storefront-deployment-update.sh set $OCIR_STOREFRONT_LOCATION $OBJECT_STORAGE_NAMESPACE $OCIR_STOREFRONT_NAME
